Upgrade base duplicati package for Ubuntu 24.04
- Old version fails to install
- https://github.com/duplicati/duplicati/issues/5178

$ cat /etc/issue
Ubuntu 24.04 LTS \n \l
$ sudo apt-get install duplicati_2.0.7.1-1_all.deb
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
Note, selecting 'duplicati' instead of 'duplicati_2.0.7.1-1_all.deb'
Some packages could not be installed. This may mean that you have
requested an impossible situation or if you are using the unstable
distribution that some required packages have not yet been created
or been moved out of Incoming.
The following information may help to resolve the situation:

The following packages have unmet dependencies:
duplicati : Depends: libappindicator0.1-cil but it is not installable or
libappindicator3-0.1-cil but it is not installable or
libayatana-appindicator1 but it is not installable
Depends: gtk-sharp2 but it is not installable
E: Unable to correct problems, you have held broken packages.